:books: [cppcon-2018] State Machines Battelfield - Naive vs STL vs Boost
Problem:
- Results from connection benchmarks are missing.

Solution:
- Add connection benchmark results.
