+++ benchmark/connection/compilation_time[s]_debug.dat
@@ -1,0 +1,10 @@
+"naive if else" 0.059094891 
+"naive switch" 0.051649451 
+"naive state pattern" 0.280232583 
+"stl variant" 0.431277559 
+"stl coroutine loop" 0.291664381 
+"stl coroutine goto" 0.305196257 
+"stl coroutine fun variant" 0.498294969 
+"boost statechart" 0.979645741 
+"boost msm" 4.688383483 
+"boost sml" 0.217062032
