+++ benchmark/connection/naive_if_else.cpp
@@ -1,65 +1,131 @@
-#include <cstdio>
-
 struct connect{};
 struct established{};
 struct ping{};
 struct disconnect{};
 struct timeout{};
 
-const auto establish = []{ std::puts("establish"); };
-const auto close = []{ std::puts("close"); };
-const auto is_valid = [](const auto&) { return true; };
-const auto resetTimeout = [] { std::puts("resetTimeout"); };
+#if defined(TEST_PERF) or defined(TEST_GBENCH)
+  static void clobber() { asm volatile("" : : : "memory"); }
+  const auto establish = []{ clobber(); };
+  const auto close = []{ clobber(); };
+  const auto is_valid = [](auto const&) { clobber(); return true; };
+  const auto reset_timeout = [] { clobber(); };
+#else
+  #include <cstdio>
+  constexpr auto establish = []{ std::puts("establish"); };
+  constexpr auto close = []{ std::puts("close"); };
+  constexpr auto is_valid = [](auto const&) { return true; };
+  constexpr auto reset_timeout = []{ std::puts("reset_timeout"); };
+#endif
 
 class Connection {
-    bool disconnected = true, connected = false, connecting = false;
+  bool disconnected = true, connected = false, connecting = false;
 
 public:
-    void process_event(const connect& event) {
-      if (disconnected or connected) {
-        establish();
-        connected = disconnected = false;
-        connecting = true;
-      }
+  constexpr void process_event(connect const&) {
+    if (disconnected) {
+      establish();
+      connected = disconnected = false;
+      connecting = true;
     }
+  }
 
-    void process_event(const disconnect& event) {
-      if (connecting or connected) {
-        close();
-        connected = connecting = false;
-        disconnected = true;
-      }
+  constexpr void process_event(disconnect const&) {
+    if (connecting or connected) {
+      close();
+      connected = connecting = false;
+      disconnected = true;
     }
+  }
 
-    void process_event(const established& event) {
-      if (connecting) {
-        connecting = disconnected = false;
-        connected = true;
-      }
+  void process_event(established const& event) {
+    if (connecting) {
+      connecting = disconnected = false;
+      connected = true;
     }
+  }
 
-    void process_event(const ping& event) {
-      if (connected and is_valid(event)) {
-        resetTimeout();
-      }
+  constexpr void process_event(ping const& event) {
+    if (connected and is_valid(event)) {
+      reset_timeout();
     }
+  }
 
-    void process_event(const timeout& event) {
-      if (connected) {
-        establish();
-        connecting = true;
-        connected = disconnected = false;
-      }
+  constexpr void process_event(timeout const&) {
+    if (connected) {
+      establish();
+      connecting = true;
+      connected = disconnected = false;
     }
+  }
 };
 
-int main() {
-  Connection connection;
-  connection.process_event(connect{});
-  connection.process_event(established{});
-  connection.process_event(ping{});
-  connection.process_event(disconnect{});
-  connection.process_event(connect{});
-  connection.process_event(established{});
-  connection.process_event(ping{});
-}
+#if defined(TEST_ASM)
+  int main() {
+    Connection connection{};
+    connection.process_event(connect{});
+    connection.process_event(established{});
+    connection.process_event(ping{});
+    connection.process_event(disconnect{});
+    connection.process_event(connect{});
+    connection.process_event(established{});
+    connection.process_event(ping{});
+  }
+#elif defined(TEST_ASM_GENERIC)
+  #include <cstdlib>
+
+  int main() {
+    Connection connection{};
+
+    while(true) {
+      switch(rand() % 5) {
+        default: break;
+        case 0: connection.process_event(connect{}); break;
+        case 1: connection.process_event(established{}); break;
+        case 2: connection.process_event(ping{}); break;
+        case 3: connection.process_event(disconnect{}); break;
+        case 4: connection.process_event(timeout{}); break;
+      }
+    }
+  }
+#elif defined(TEST_PERF)
+  int main() {
+    Connection connection{};
+
+    for (auto i = 0; i < 100'000'000; ++i) {
+      connection.process_event(connect{});
+      connection.process_event(established{});
+      connection.process_event(ping{});
+      connection.process_event(disconnect{});
+    }
+  }
+#elif defined(TEST_GBENCH)
+  #include <benchmark/benchmark.h>
+
+  static void BM_naive_if_else(benchmark::State& state) {
+    constexpr auto size = 1'000'000;
+
+    int dispatch[size]{};
+    for (auto i = 0; i < size; ++i) {
+      dispatch[i] = rand() % 4;
+    }
+
+    Connection connection;
+
+    auto i = 0;
+    for (auto _ : state) {
+      switch(dispatch[i++]) {
+        default: break;
+        case 0: connection.process_event(connect{}); break;
+        case 1: connection.process_event(established{}); break;
+        case 2: connection.process_event(ping{}); break;
+        case 3: connection.process_event(disconnect{}); break;
+        case 4: connection.process_event(timeout{}); break;
+      }
+      if (i++ >= size) i = {};
+    }
+  }
+
+  BENCHMARK(BM_naive_if_else);
+  BENCHMARK_MAIN();
+#endif
