+++ benchmark/connection/naive_state_pattern.cpp
@@ -1,8 +1,8 @@
-struct connect{};
-struct established{};
-struct ping{};
-struct disconnect{};
-struct timeout{};
+struct connect {};
+struct ping {};
+struct established {};
+struct timeout {};
+struct disconnect {};
 
 #if defined(TEST_PERF) or defined(TEST_GBENCH)
   static void clobber() { asm volatile("" : : : "memory"); }
@@ -18,28 +18,94 @@ struct timeout{};
   constexpr auto reset_timeout = []{ std::puts("reset_timeout"); };
 #endif
 
-#include <boost/sml.hpp>
+#include <memory>
 
-namespace sml = boost::sml;
-
-struct connection {
-  auto operator()() const {
-    using namespace sml;
-    return make_transition_table(
-      * "Disconnected"_s + event<connect> / establish                = "Connecting"_s,
-        "Connecting"_s   + event<established>                        = "Connected"_s,
-        "Connected"_s    + event<ping> [ is_valid ] / reset_timeout,
-        "Connected"_s    + event<timeout> / establish                = "Connecting"_s,
-        "Connected"_s    + event<disconnect> / close                 = "Disconnected"_s
-    );
-  }
+struct State {
+  virtual ~State() noexcept = default;
+  virtual void process_event(connect const&) { }
+  virtual void process_event(ping const&) { }
+  virtual void process_event(established const&) { }
+  virtual void process_event(timeout const&) { }
+  virtual void process_event(disconnect const&) { }
 };
 
-using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_expr>>;
+class Connecting;
+class Connected;
+class Disconnected;
+
+class Connection {
+ public:
+  template<class TState>
+  void init_state() { change_state<TState>(); }
+
+  template<class TEvent>
+  void process_event(TEvent const& event) {
+    state->process_event(event);
+  }
+
+  template<class TState>
+  void change_state() {
+    state = std::make_unique<TState>(*this);
+  }
+
+ private:
+  std::unique_ptr<State> state{};
+};
+
+class Disconnected : public State {
+ public:
+  explicit Disconnected(Connection& connection) : connection{connection} {}
+
+  void process_event(connect const&) override final {
+    establish();
+    connection.change_state<Connecting>();
+  }
+
+ private:
+  Connection& connection;
+};
+
+class Connecting : public State {
+ public:
+  explicit Connecting(Connection& connection) : connection{connection} {}
+
+  void process_event(established const&) override final {
+    connection.change_state<Connected>();
+  }
+
+ private:
+  Connection& connection;
+};
+
+class Connected : public State {
+ public:
+  explicit Connected(Connection& connection) : connection{connection} {}
+
+  void process_event(ping const& event) override final {
+    if (is_valid(event)) {
+      reset_timeout();
+    }
+  }
+
+  void process_event(timeout const&) override final {
+    establish();
+    connection.change_state<Connecting>();
+  }
+
+  void process_event(disconnect const&) override final {
+    close();
+    connection.change_state<Disconnected>();
+  }
+
+ private:
+  Connection& connection;
+};
 
 #if defined(TEST_ASM)
   int main() {
     Connection connection{};
+    connection.init_state<Disconnected>();
+
     connection.process_event(connect{});
     connection.process_event(established{});
     connection.process_event(ping{});
@@ -53,6 +119,7 @@ using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_e
 
   int main() {
     Connection connection{};
+    connection.init_state<Disconnected>();
 
     while(true) {
       switch(rand() % 5) {
@@ -68,6 +135,7 @@ using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_e
 #elif defined(TEST_PERF)
   int main() {
     Connection connection{};
+    connection.init_state<Disconnected>();
 
     for (auto i = 0; i < 100'000'000; ++i) {
       connection.process_event(connect{});
@@ -79,7 +147,7 @@ using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_e
 #elif defined(TEST_GBENCH)
   #include <benchmark/benchmark.h>
 
-  static void BM_boost_sml(benchmark::State& state) {
+  static void BM_naive_state_pattern_v1(benchmark::State& state) {
     constexpr auto size = 1'000'000;
 
     int dispatch[size]{};
@@ -88,6 +156,7 @@ using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_e
     }
 
     Connection connection;
+    connection.init_state<Disconnected>();
 
     auto i = 0;
     for (auto _ : state) {
@@ -103,6 +172,6 @@ using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_e
     }
   }
 
-  BENCHMARK(BM_boost_sml);
+  BENCHMARK(BM_naive_state_pattern_v1);
   BENCHMARK_MAIN();
 #endif
