+++ benchmark/connection/naive_state_pattern_v2.cpp
@@ -1,8 +1,8 @@
-struct connect{};
-struct established{};
-struct ping{};
-struct disconnect{};
-struct timeout{};
+struct connect {};
+struct ping {};
+struct established {};
+struct timeout {};
+struct disconnect {};
 
 #if defined(TEST_PERF) or defined(TEST_GBENCH)
   static void clobber() { asm volatile("" : : : "memory"); }
@@ -18,24 +18,67 @@ struct timeout{};
   constexpr auto reset_timeout = []{ std::puts("reset_timeout"); };
 #endif
 
-#include <boost/sml.hpp>
+#include <memory>
 
-namespace sml = boost::sml;
-
-struct connection {
-  auto operator()() const {
-    using namespace sml;
-    return make_transition_table(
-      * "Disconnected"_s + event<connect> / establish                = "Connecting"_s,
-        "Connecting"_s   + event<established>                        = "Connected"_s,
-        "Connected"_s    + event<ping> [ is_valid ] / reset_timeout,
-        "Connected"_s    + event<timeout> / establish                = "Connecting"_s,
-        "Connected"_s    + event<disconnect> / close                 = "Disconnected"_s
-    );
-  }
+struct State {
+  virtual ~State() noexcept = default;
+  virtual std::unique_ptr<State> process_event(connect const&) { return {}; }
+  virtual std::unique_ptr<State> process_event(ping const&) { return {}; }
+  virtual std::unique_ptr<State> process_event(established const&) { return {}; }
+  virtual std::unique_ptr<State> process_event(timeout const&) { return {}; }
+  virtual std::unique_ptr<State> process_event(disconnect const&) { return {}; }
 };
 
-using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_expr>>;
+struct Disconnected : State {
+  std::unique_ptr<State> process_event(connect const&) override final;
+};
+
+struct Connecting : State {
+  std::unique_ptr<State> process_event(established const&) override final;
+};
+
+struct Connected : State {
+  std::unique_ptr<State> process_event(ping const& event) override final;
+  std::unique_ptr<State> process_event(timeout const&) override final;
+  std::unique_ptr<State> process_event(disconnect const&) override final;
+};
+
+std::unique_ptr<State> Disconnected::process_event(connect const&) {
+  establish();
+  return std::make_unique<Connecting>();
+}
+
+std::unique_ptr<State> Connecting::process_event(established const&) {
+  return std::make_unique<Connected>();
+}
+
+std::unique_ptr<State> Connected::process_event(ping const& event) {
+  if (is_valid(event)) {
+    reset_timeout();
+  }
+  return {};
+}
+
+std::unique_ptr<State> Connected::process_event(timeout const&) {
+  establish();
+  return std::make_unique<Connecting>();
+}
+
+std::unique_ptr<State> Connected::process_event(disconnect const&) {
+  close();
+  return std::make_unique<Disconnected>();
+}
+
+struct Connection {
+  template<class TEvent>
+  void process_event(TEvent const& event) {
+    if (auto new_state = state->process_event(event); new_state) {
+      state = std::move(new_state);
+    }
+  }
+
+  std::unique_ptr<State> state = std::make_unique<Disconnected>();
+};
 
 #if defined(TEST_ASM)
   int main() {
@@ -79,7 +122,7 @@ using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_e
 #elif defined(TEST_GBENCH)
   #include <benchmark/benchmark.h>
 
-  static void BM_boost_sml(benchmark::State& state) {
+  static void BM_naive_state_pattern_v2(benchmark::State& state) {
     constexpr auto size = 1'000'000;
 
     int dispatch[size]{};
@@ -103,6 +146,6 @@ using Connection = sml::sm<connection, sml::dispatch<sml::back::policies::fold_e
     }
   }
 
-  BENCHMARK(BM_boost_sml);
+  BENCHMARK(BM_naive_state_pattern_v2);
   BENCHMARK_MAIN();
 #endif
