+++ benchmark/connection/performance_branch-misses_release.dat
@@ -1,0 +1,10 @@
+"naive if else" 8,528      
+"naive switch" 8,614      
+"naive state pattern" 200,012,399      
+"stl variant" 9,034      
+"stl coroutine loop" 13,318      
+"stl coroutine goto" 12,988      
+"stl coroutine fun variant" 17,247      
+"boost statechart" 31,202      
+"boost msm" 8,452      
+"boost sml" 8,493
