+++ benchmark/connection/performance_branches_release.dat
@@ -1,0 +1,10 @@
+"naive if else" 100,254,967      
+"naive switch" 100,254,938      
+"naive state pattern" 12,900,277,090      
+"stl variant" 1,100,255,178      
+"stl coroutine loop" 2,900,271,998      
+"stl coroutine goto" 2,700,271,955      
+"stl coroutine fun variant" 14,400,276,476      
+"boost statechart" 23,300,283,166      
+"boost msm" 900,255,655      
+"boost sml" 100,254,915
