+++ benchmark/connection/performance_cycles_release.dat
@@ -1,0 +1,10 @@
+"naive if else" 101,226,047      
+"naive switch" 101,315,085      
+"naive state pattern" 22,185,310,766      
+"stl variant" 1,628,309,122      
+"stl coroutine loop" 7,687,983,920      
+"stl coroutine goto" 7,623,830,065      
+"stl coroutine fun variant" 20,799,234,095      
+"boost statechart" 39,881,690,826      
+"boost msm" 2,202,617,009      
+"boost sml" 101,255,568
