+++ benchmark/connection/performance_time[s]_release.dat
@@ -1,0 +1,10 @@
+"naive if else" 0.057759532 
+"naive switch" 0.039429509 
+"naive state pattern" 6.774054670 
+"stl variant" 0.583537661 
+"stl coroutine loop" 2.405965772 
+"stl coroutine goto" 2.369846175 
+"stl coroutine fun variant" 6.400204488 
+"boost statechart" 12.241939420 
+"boost msm" 0.784781340 
+"boost sml" 0.062128123
